# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API บริการหลังบ้านสำหรับจัดการบันทึกรายรับรายจ่ายและวิเคราะห์สลิปด้วย Gemini AI",
        "title": "CashLog API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "jaruvat.see@gmail.com"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/accounts": {
            "get": {
                "description": "ดึงรายการบัญชีทั้งหมดที่ is_active=true",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "ดึงรายการ Account ที่ใช้งานอยู่",
                "responses": {
                    "200": {
                        "description": "FetchActiveAccounts successful \u003cbr\u003emessage: FetchActiveAccounts successfull",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jaruvat303_cashlog_pkg_response.JsonResponse-array_github_com_Jaruvat303_cashlog_internal_delivery_http_v1_dto_AccountResponse"
                        }
                    },
                    "499": {
                        "description": "Client Closed Request  \u003cbr\u003eerror_code: REQUEST_CANCELED \u003cbr\u003emessage: The request was canceled by the user",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jaruvat303_cashlog_internal_delivery_http_v1_dto.ErrorResponseDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error \u003cbr\u003eerror_code: INTERNAL_SERVER_ERROR or INTERNAL_DATABASE_ERROR \u003cbr\u003emessage: Something went wrong, please try again later",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jaruvat303_cashlog_internal_delivery_http_v1_dto.ErrorResponseDTO"
                        }
                    },
                    "504": {
                        "description": "Gateway Timeout \u003cbr\u003eerror_code: DATABASE_TIMEOUT \u003cbr\u003emessage: The database operation timed out, please try again",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jaruvat303_cashlog_internal_delivery_http_v1_dto.ErrorResponseDTO"
                        }
                    }
                }
            },
            "post": {
                "description": "สร้างบัญชีใหม่ โดยต้องระบุชื่อและประเภทของบัญชี",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "สร้าง Account ใหม่",
                "parameters": [
                    {
                        "description": "ข้อมูลสำหรับสร้าง Account",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Jaruvat303_cashlog_internal_delivery_http_v1_dto.CreateAccountInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "create account successfull",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jaruvat303_cashlog_pkg_response.JsonResponse-github_com_Jaruvat303_cashlog_internal_delivery_http_v1_dto_AccountResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request  \u003cbr\u003eerror_code: INVALID_INPUT_PARAMETERS \u003cbr\u003emessage: 1. Invalid JSON format.  2. Validation failed for the request data.",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jaruvat303_cashlog_internal_delivery_http_v1_dto.ErrorResponseDTO"
                        }
                    },
                    "499": {
                        "description": "Client Closed Request  \u003cbr\u003eerror_code: REQUEST_CANCELED \u003cbr\u003emessage: The request was canceled by the user",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jaruvat303_cashlog_internal_delivery_http_v1_dto.ErrorResponseDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error \u003cbr\u003eerror_code: INTERNAL_SERVER_ERROR or INTERNAL_DATABASE_ERROR \u003cbr\u003emessage: Something went wrong, please try again later",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jaruvat303_cashlog_internal_delivery_http_v1_dto.ErrorResponseDTO"
                        }
                    },
                    "504": {
                        "description": "Gateway Timeout \u003cbr\u003eerror_code: DATABASE_TIMEOUT \u003cbr\u003emessage: The database operation timed out, please try again",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jaruvat303_cashlog_internal_delivery_http_v1_dto.ErrorResponseDTO"
                        }
                    }
                }
            }
        },
        "/accounts/{id}": {
            "delete": {
                "description": "ปิดใช้งานบัญชี (soft delete) โดยต้องระบุ ID ของบัญชีที่ต้องการลบ",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "ลบ Account",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID ของ Account ที่ต้องการลบ",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request \u003cbr\u003eerror_code: INVALID_INPUT_PARAMETERS \u003cbr\u003eMessage: Invalid ID format. The path parameter 'id' must be a positive integer.",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jaruvat303_cashlog_internal_delivery_http_v1_dto.ErrorResponseDTO"
                        }
                    },
                    "499": {
                        "description": "Client Closed Request  \u003cbr\u003eerror_code: REQUEST_CANCELED \u003cbr\u003emessage: The request was canceled by the user",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jaruvat303_cashlog_internal_delivery_http_v1_dto.ErrorResponseDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error \u003cbr\u003eerror_code: INTERNAL_SERVER_ERROR or INTERNAL_DATABASE_ERROR \u003cbr\u003emessage: Something went wrong, please try again later",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jaruvat303_cashlog_internal_delivery_http_v1_dto.ErrorResponseDTO"
                        }
                    },
                    "504": {
                        "description": "Gateway Timeout \u003cbr\u003eerror_code: DATABASE_TIMEOUT \u003cbr\u003emessage: The database operation timed out, please try again",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jaruvat303_cashlog_internal_delivery_http_v1_dto.ErrorResponseDTO"
                        }
                    }
                }
            },
            "patch": {
                "description": "อัปเดตข้อมูลบัญชี โดยต้องระบุ ID ของบัญชีที่ต้องการอัปเดต",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "อัปเดต Account",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID ของ Account ที่ต้องการอัปเดต",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ข้อมูลสำหรับอัปเดต Account",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Jaruvat303_cashlog_internal_delivery_http_v1_dto.UpdateAccountInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "update account successfull",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jaruvat303_cashlog_pkg_response.JsonResponse-github_com_Jaruvat303_cashlog_internal_delivery_http_v1_dto_AccountResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request  \u003cbr\u003eerror_code: INVALID_INPUT_PARAMETERS \u003cbr\u003emessage: 1. Invalid ID format (must be a positive integer) 2. Invalid JSON format 3. Validation failed for the request data",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jaruvat303_cashlog_internal_delivery_http_v1_dto.ErrorResponseDTO"
                        }
                    },
                    "499": {
                        "description": "Client Closed Request  \u003cbr\u003eerror_code: REQUEST_CANCELED \u003cbr\u003emessage: The request was canceled by the user",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jaruvat303_cashlog_internal_delivery_http_v1_dto.ErrorResponseDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error \u003cbr\u003eerror_code: INTERNAL_SERVER_ERROR or INTERNAL_DATABASE_ERROR \u003cbr\u003emessage: Something went wrong, please try again later",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jaruvat303_cashlog_internal_delivery_http_v1_dto.ErrorResponseDTO"
                        }
                    },
                    "504": {
                        "description": "Gateway Timeout \u003cbr\u003eerror_code: DATABASE_TIMEOUT \u003cbr\u003emessage: The database operation timed out, please try again",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jaruvat303_cashlog_internal_delivery_http_v1_dto.ErrorResponseDTO"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_Jaruvat303_cashlog_internal_delivery_http_v1_dto.AccountResponse": {
            "type": "object",
            "properties": {
                "account_type": {
                    "type": "string",
                    "example": "bank"
                },
                "color_hex": {
                    "type": "string",
                    "example": "#4F46E5"
                },
                "icon_key": {
                    "type": "string",
                    "example": "bank"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "matching_keywords": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "SCB"
                    ]
                },
                "name": {
                    "type": "string",
                    "example": "SCB"
                },
                "opening_balance": {
                    "type": "number",
                    "example": 1000
                }
            }
        },
...
        "github_com_Jaruvat303_cashlog_internal_delivery_http_v1_dto.CreateAccountInput": {
            "type": "object",
            "required": [
                "account_type",
                "matching_keywords",
                "name"
            ],
            "properties": {
                "account_type": {
                    "type": "string",
                    "enum": [
                        "cash",
                        "bank",
                        "investment",
                        "ewallet"
                    ],
                    "example": "bank"
                },
                "color_hex": {
                    "type": "string",
                    "maxLength": 10,
                    "example": "#4F46E5"
                },
                "icon_key": {
                    "type": "string",
                    "maxLength": 50,
                    "example": "bank"
                },
                "matching_keywords": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "SCB"
                    ]
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "SCB"
                },
                "opening_balance": {
                    "type": "number",
                    "minimum": 0,
                    "example": 1000
                }
            }
        },
...
        "github_com_Jaruvat303_cashlog_internal_delivery_http_v1_dto.ErrorResponseDTO": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "string",
                    "example": "ERROR_CODE_HERE"
                },
                "message": {
                    "type": "string",
                    "example": "ข้อความอธิบายความผิดพลาดตามสถานการณ์"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "github_com_Jaruvat303_cashlog_internal_delivery_http_v1_dto.UpdateAccountInput": {
            "type": "object",
            "required": [
                "matching_keywords"
            ],
            "properties": {
                "account_type": {
                    "type": "string",
                    "enum": [
                        "cash",
                        "bank",
                        "investment",
                        "ewallet"
                    ],
                    "example": "bank"
                },
                "color_hex": {
                    "type": "string",
                    "maxLength": 10,
                    "example": "#4F46E5"
                },
                "icon_key": {
                    "type": "string",
                    "maxLength": 50,
                    "example": "bank"
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "matching_keywords": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "SCB"
                    ]
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "SCB"
                }
            }
        },
...
        "github_com_Jaruvat303_cashlog_pkg_response.JsonResponse-array_github_com_Jaruvat303_cashlog_internal_delivery_http_v1_dto_AccountResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_Jaruvat303_cashlog_internal_delivery_http_v1_dto.AccountResponse"
                    }
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "github_com_Jaruvat303_cashlog_pkg_response.JsonResponse-github_com_Jaruvat303_cashlog_internal_delivery_http_v1_dto_AccountResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_Jaruvat303_cashlog_internal_delivery_http_v1_dto.AccountResponse"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    }
}